Added API for user-defined code to Prioritizer, changed crawl priority from Int to Long

+++ src/main/scala/abwcf/data/Page.scala
@@ -6,12 +6,12 @@ enum PageStatus {
 
 /**
  * A page that is currently being crawled or has already been crawled.
- * 
+ *
  * [[Page]]s are persisted in the database.
  */
 case class Page(
                  url: String,
                  status: PageStatus,
                  crawlDepth: Int,
-                 crawlPriority: Int
+                 crawlPriority: Long
                )
